Work on exp-4340: Add url adclick telemetry event check.
